Fixed the grid square to zone lookup -- our solution is to replace st_within with st_intersects. It jsut works for datazone 11 and we tested it does the same thing as st_within when we use dots in polygons with the other datazone and lsoa files.
NOTE:: we don't use st_join because its more complicated -- some zones (i.e. lsoa) lack grid centroids so we have to return to the zone list and find the nearest.

Our previous code was not wrong I was jsut confused and have added more comments

Now to continue with the rest of analysis.