minor change in title to be consistent with segregation
Terminology can easily be confusing in this field. It would help the paper to follow Massey and Denton’s approach: use ‘segregation’ to cover all the various dimensions; use (un)evenness in reference to e.g. dissimilarity; use centralisation in relation to RCI. See for example title to Figure 2 and text page 19, line 48